Third milestone (#16)
* Updating third-milestone branch (#13)

* Removed disk related functionalities

* Equal instead of Compare

* Test change

* Addition of disk functionalities

* Addition of disk functionalities in pool.go and frame.go

* Partial read error addition

* Read/IO implementation

* Presentation outline

* Presentation outline removal

* Add node back to frame when doing IO

* Add function

* Keyed litteral addition

* FrameOverflowError

* Bug correction

* InvalidNodeError addition

* Check if node is valid inside memory middleware

* Metadata marshal/unmarshal and test

* Read/Write to disk

* InvalidMetadataError addition

* Correction in frame init

* Metadata simplification

* Read/Write to disk

* Removed debug primitives

* Removed redundant logs

* Anonymous function instead of global one

* Test improvement

* Typo correction

* Page not dirty when loading from disk

* To each their own file

* Better error handling

* BufferpoolLimitError

* Better error handling when registering

* HBTrie page eviction test

* metaSize for Trie

* Update root on operations

* Size assertion

* Marshal boolean value

* Setter/getter aiddition

* Field 'cursor' addition

* First draft of Write/Read trie

* Addition of errors

* Tests revisited and refactoring

* Addition of tests

* Entry test extension

* Compare functionality

* Compare functionality

* Addition of errors

* Extensive testing with multiple disk write/retrieve

* Correction of minor bugs

* Increase of pool limit

* return number of frames on read

* Enhanced test (still not pass)

* Addition of errors

* Changed test

* To each frame their own file

* To each frame their own file

* :bug: Finally nasty bug is fixed. Page id of the root was not updated after split and pageId was returned as frameId during reading trie process.

* use userdefined filepath for database files.

* Write buffer index has been implemented. Uses golang map as hashtable.

* Removed unnecessary debug variable 'good'

* Removed commented functions

* Comments addition and cleanup

* Test correction

* README refactoring

* Refactorage for submission

+++ internal/pool/page.go
@@ -1,6 +1,6 @@
 package pool
 
-var PageSize uint16 = 4096
+var PageSize uint64 = 4096
 
 // Page is the unit of the Bufferpool
 
@@ -20,6 +20,7 @@ type Page struct {
 
 }
 
+// Initialises a new page with given id. On initialisation the dirty flag is set to true.
 func NewPage(id uint64) *Page {
 	return &Page{Id: id, Dirty: true, prev: nil, next: nil}
 }
